+++ CHANGELOG.md
@@ -3,6 +3,41 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.30.8](https://github.com/IBM/carbon-charts/compare/v0.30.7...v0.30.8) (2020-04-03)
+
+**Note:** Version bump only for package @carbon/charts-monorepo
+
+
+
+
+
+## [0.30.7](https://github.com/IBM/carbon-charts/compare/v0.30.6...v0.30.7) (2020-03-31)
+
+
+### Bug Fixes
+
+* **core:** update import path carbon-components ([e9fca73](https://github.com/IBM/carbon-charts/commit/e9fca73cd3768b1b85247a2798908c26da99216b)), closes [#546](https://github.com/IBM/carbon-charts/issues/546)
+
+
+
+
+
+## [0.30.6](https://github.com/IBM/carbon-charts/compare/v0.30.5...v0.30.6) (2020-03-30)
+
+**Note:** Version bump only for package @carbon/charts-monorepo
+
+
+
+
+
+## [0.30.5](https://github.com/IBM/carbon-charts/compare/v0.30.4...v0.30.5) (2020-03-30)
+
+**Note:** Version bump only for package @carbon/charts-monorepo
+
+
+
+
+
 ## [0.30.4](https://github.com/IBM/carbon-charts/compare/v0.30.3...v0.30.4) (2020-03-27)
 
 **Note:** Version bump only for package @carbon/charts-monorepo